+++ src/registration-form/index.js
@@ -3,8 +3,8 @@ import Validator from '../Validator';
 import H1 from '../components/H1';
 import Field from '../components/Field';
 import Label from '../components/Lable';
-import FieldLabel from '../components/FieldLabel';
-import FieldHint from '../components/FieldHint';
+import FieldLabel from '../components/Field-Label';
+import FieldHint from '../components/Field-Hint';
 import Input from '../components/Input';
 import InputPassword from '../components/Input-Password';
 import Submit from '../components/Submit';
